Moved theme to style.js
Moved theming to style.js following @vfcheung feedback.
