+++ src/style.js
@@ -1,11 +1,46 @@
 import { StyleSheet } from 'react-native';
+import { DefaultTheme, configureFonts } from 'react-native-paper';
 
-const styles = (props) => StyleSheet.create({
+const fontConfig = {
+  default: {
+    regular: {
+      fontFamily: 'sans-serif',
+      fontWeight: 'normal',
+    },
+    medium: {
+      fontFamily: 'sans-serif-medium',
+      fontWeight: 'normal',
+    },
+    light: {
+      fontFamily: 'sans-serif-light',
+      fontWeight: 'normal',
+    },
+    thin: {
+      fontFamily: 'sans-serif-thin',
+      fontWeight: 'normal',
+    },
+  },
+};
+
+export const theme = {
+  ...DefaultTheme,
+  roundness: 20,
+  fonts: configureFonts(fontConfig),
+  colors: {
+    ...DefaultTheme.colors,
+    primary: '#1F7FC0',
+    accent: '#8EAEC3',
+    background: '#E6EFF5',
+    text: '#194261',
+  },
+};
+
+const styles = StyleSheet.create({
   container: {
     flex: 1,
     justifyContent: 'center',
     alignItems: 'center',
-    backgroundColor: props.theme.colors.background,
+    backgroundColor: theme.colors.background,
   },
   textInput: {
     height: 40,
